requirements.txt
numpy==1.24.3
opencv-python==4.8.0
openalpr-py==1.0.0